backend/requirements.txt
channels==4.2.0
crewai==0.95.0
distro==1.9.0
dj-rest-auth==7.0.1
Django==5.1.4
django-allauth==65.3.1
django-cors-headers==4.6.0
django-extensions==3.2.3
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
holidays==0.63
kubernetes==31.0.0
openai==1.59.3
pillow==10.4.0
requests==2.32.3
setuptools==75.6.0
uvicorn==0.34.0
crawl4ai~=0.4.247
websockets~=14.1
asgiref~=3.8.1
PyJWT~=2.10.1
six~=1.17.0
python-dotenv~=1.0.1
protobuf~=5.29.3
pydantic~=2.10.5
psycopg2
watchgod